+++ Codebase/barcode_app/Changelog.md
@@ -1,1 +1,2 @@
 1. create django project with docker build script
+2. create a working prototype for the barcode scanner